requirements.txt
aiogram==3.19.0
asyncpg==0.30.0
openai==1.72.0
tiktoken==0.9.0
aiofiles==24.1.0
Pillow==11.1.0
opencv-python==4.11.0.86
numpy==2.2.4
google-cloud-vision==3.10.1
google-generativeai==0.8.4
requests==2.32.3
tika==3.1.0
python-dotenv==1.1.0
protobuf==5.29.4